+++ myMovies/nbproject/project.properties
@@ -1,9 +1,10 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
+application.title=myMovies
+application.vendor=dinob
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -30,6 +31,7 @@ dist.archive.excludes=
 dist.dir=dist
 dist.jar=${dist.dir}/myMovies.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+endorsed.classpath=
 excludes=
 file.reference.gson-2.8.5.jar-1=libs\\gson-2.8.5.jar
 includes=**
