--- myMovies/src/database/DBManager.java
@@ -1,12 +1,0 @@
-/*
- * Created by Batzonis Constantinos - Klianis Christos - Servozlidis Giwrgos
- */
-package database;
-
-/**
- *
- * @author dinob
- */
-public class DBManager {
-    
-}
