+++ myMovies/src/design/AddNewList.form
