+++ myMovies/src/design/MyFavoriteList.form
